Implement standalone mode
