+++ src/config.ts
@@ -1,0 +1,2 @@
+const standalone = !!import.meta.env.VITE_RR_STANDALONE;
+export { standalone };
